% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_analysis.R
\name{analyse}
\alias{analyse}
\title{analyse Function}
\usage{
analyse(
  corrKey = "N-N",
  T = time,
  n_src = 16,
  n_conf = 1224,
  path_letter = "b",
  gi = "Gi_Cg5",
  gf = "Gf_Cg5",
  step = 24,
  mom_tag = "px00py00pz00",
  conf_start = 0
)
}
\description{
This function allows you to read a specific correlator from an aff-file.
}
\examples{
cf <- analyze()
}
