% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_analysis.R
\name{plot.2combs}
\alias{plot.2combs}
\title{plot.2comb Function}
\usage{
\method{plot}{`2combs`}(
  gi1 = "Gi_Cg5",
  gf1 = "Gf_Cg5",
  gi2 = "Gi_Cg5gt",
  gf2 = "Gf_Cg5gt",
  mom.tag = "px00py00pz00"
)
}
\description{
This function allows you to compare two specific correlator ratios.
}
\examples{

}
