% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_analysis.R
\name{plot.ratio}
\alias{plot.ratio}
\title{plot.ratio Function}
\usage{
\method{plot}{ratio}(c2, c3, t1, t2, add = FALSE)
}
\description{
This function allows you to plot the ratio of two correlator functions.
}
\examples{

}
