% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_analysis.R
\name{plot_2combs}
\alias{plot_2combs}
\title{plot Function to compare 2 ratios}
\usage{
plot_2combs(
  gi1 = "Gi_Cg5",
  gf1 = "Gf_Cg5",
  gi2 = "Gi_Cg5gt",
  gf2 = "Gf_Cg5gt",
  mom.tag = "px00py00pz00",
  col2 = "blue"
)
}
\description{
This function allows you to compare two specific correlator ratios.
}
