% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_analysis.R
\name{plot_comb}
\alias{plot_comb}
\title{plot Function for single combination}
\usage{
plot_comb(g.i, g.f, mom.tag, bool)
}
\description{
This function allows you to plot one specific correlator (2pt and 3pt) combination.
}
