% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_analysis.R
\name{plot_ratio}
\alias{plot_ratio}
\title{plot_ratio Function}
\usage{
plot_ratio(c2, c3, t1, t2, add = FALSE, .col = "black")
}
\description{
This function allows you to plot the ratio of two correlator functions.
}
\examples{

}
